pyproject.toml
[project]
name = "narrative-graph"
version = "0.1.0"
description = "Narrative Graph Intelligence MVP - A mini Constellation-like system for narrative analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer>=0.9.0",
    "pyyaml>=6.0.1",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
    
    # Data processing
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "sqlite-utils>=3.35.0",
    
    # NLP
    "spacy>=3.7.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "hdbscan>=0.8.33",
    "langdetect>=1.0.9",
    
    # Graph
    "neo4j>=5.14.0",
    
    # UI/API
    "streamlit>=1.29.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "plotly>=5.18.0",
    "pyvis>=0.3.2",
    
    # LLM (optional)
    "openai>=1.3.0",
    
    # Utilities
    "httpx>=0.25.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "testcontainers>=3.7.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
narrative-graph = "narrative_graph.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/narrative_graph"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
